Finish UserRegistration and ManagerGeneration components.

+++ src/main/java/yorku/eecs3311/ParkingRate.java
@@ -1,18 +1,18 @@
 package yorku.eecs3311;
 
 public enum ParkingRate {
-	STUDENT(5),
-	FACULTY(8),
-	STAFF(10),
-	VISITOR(15);
+	STUDENT(5.0),
+	FACULTY(8.0),
+	STAFF(10.0),
+	VISITOR(15.0);
 	
-	private int rate;
+	private double rate;
 
-	ParkingRate(int rate) {
+	ParkingRate(double rate) {
 		this.rate = rate;
 	}
 	
-	public int getValue() {
+	public double getValue() {
 		return rate;
 	}
 }
